correction

+++ cpu_2d/compression/cpusimd.hh
@@ -63,7 +63,7 @@ void CpuSimd<T, T_C>::debugCompression(T *fq, const size_t size) const
 }
 
 template <typename T, typename T_C>
-inline void CpuSimd<T, T_C>::compress(T * restrict fq_64, const int32_t size, T_C ** restrict compressed_fq_32,
+inline void CpuSimd<T, T_C>::compress(T * restrict fq_64, const size_t size, T_C ** restrict compressed_fq_32,
                                size_t &compressedsize) const
 {
     if (isCompressible(size))
